Microservicios/Microservicios/pagos/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.28.1
marshmallow==3.21.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.31.0
Flask-Cors==3.0.10